requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
whitenoise>=6.4.0
gunicorn>=20.1.0
mysqlclient>=2.1.1  # For MySQL database
pillow>=9.5.0
python-dotenv>=1.0.0
django-import-export==3.3.1
openpyxl==3.1.2
pandas>=2.0.0
xlrd>=2.0.1
django-cache-memoize>=0.1.10